requirements.txt
# Core Framework
flask>=3.0.0

# OpenAI
openai>=1.12.0

# Google APIs
google-api-python-client>=2.118.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Server
gunicorn>=21.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2024.1
